Rename package "aux" to "auxiliary" for windows users
Windows it is a name of DOS AUX device and is reserved
