+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/jms/SendToQueueFromWithinTransactionTest.java
@@ -26,9 +26,9 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.test.integration.common.jms.JMSOperations;
-import org.jboss.as.test.smoke.jms.aux.CreateQueueSetupTask;
-import org.jboss.as.test.smoke.jms.aux.QueueMessageDrivenBean;
-import org.jboss.as.test.smoke.jms.aux.TransactedQueueMessageSender;
+import org.jboss.as.test.smoke.jms.auxiliary.CreateQueueSetupTask;
+import org.jboss.as.test.smoke.jms.auxiliary.QueueMessageDrivenBean;
+import org.jboss.as.test.smoke.jms.auxiliary.TransactedQueueMessageSender;
 import org.jboss.logging.Logger;
 import org.jboss.shrinkwrap.api.ArchivePaths;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
