+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/jms/SendToTopicFromWithinTransactionTest.java
@@ -25,11 +25,10 @@ package org.jboss.as.test.smoke.jms;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.arquillian.api.ServerSetup;
-import org.jboss.as.controller.client.ModelControllerClient;
 import org.jboss.as.test.integration.common.jms.JMSOperations;
-import org.jboss.as.test.smoke.jms.aux.CreateTopicSetupTask;
-import org.jboss.as.test.smoke.jms.aux.TopicMessageDrivenBean;
-import org.jboss.as.test.smoke.jms.aux.TransactedTopicMessageSender;
+import org.jboss.as.test.smoke.jms.auxiliary.CreateTopicSetupTask;
+import org.jboss.as.test.smoke.jms.auxiliary.TopicMessageDrivenBean;
+import org.jboss.as.test.smoke.jms.auxiliary.TransactedTopicMessageSender;
 import org.jboss.logging.Logger;
 import org.jboss.shrinkwrap.api.ArchivePaths;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
