+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/jms/auxiliary/TransactedQueueMessageSender.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.smoke.jms.aux;
+package org.jboss.as.test.smoke.jms.auxiliary;
 
 import org.jboss.logging.Logger;
